adding routers and separating calls into different files
